Add hi-hat choke to drum machine; remove legacy PatternSequencer
The new DrumPatternSampleProvider (which inherits the name from the
deleted original) tracks active voices per choke group: when a note
fires whose group already has a ringing voice, the previous voice
is faded out over 10 ms via FadeInOutSampleProvider. Closed-hat
(note 2) and open-hat (note 3) share group 1, so they cut each
other the way a physical hi-hat pedal does.

The faded voice currently remains in the mixer (emitting silence)
until its underlying sample naturally ends. With short kit samples
this is fine; a proper voice manager that drops faded voices early
will land alongside the future SoundFont/sfz sampler — documented
in the architecture doc's deferred-items section.

The legacy PatternSequencer, the original DrumPatternSampleProvider,
and PatternSequencerTests are removed now that the sequencing engine
is validated. The ViewModel's "Use legacy" toggle and the matching
checkbox in the XAML come out with them.
